+++ src/services/TestTypeServices.java
@@ -6,10 +6,14 @@ import models.TestType;
 import java.sql.*;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.NoSuchElementException;
 
 public class TestTypeServices {
 
     public void createTestType(TestType testType) {
+        if (testType == null) {
+            throw new NullPointerException("The test type is null");
+        }
         String sql = "INSERT INTO TestType (testType) VALUES (?)";
         try (Connection conn = DataBaseConnection.getConnection();
              PreparedStatement pstmt = conn.prepareStatement(sql)) {
@@ -62,13 +66,20 @@ public class TestTypeServices {
     }
 
     public void updateTestType(String oldType, String newType) {
+        if (oldType == null || newType == null) {
+            throw new NullPointerException("The test type is null");
+        }
         String sql = "UPDATE TestType SET testType = ? WHERE testType = ?";
         try (Connection conn = DataBaseConnection.getConnection();
              PreparedStatement pstmt = conn.prepareStatement(sql)) {
 
             pstmt.setString(1, newType);
             pstmt.setString(2, oldType);
-            pstmt.executeUpdate();
+            int affectedrows = pstmt.executeUpdate();
+
+            if (affectedrows == 0) {
+                throw new NoSuchElementException("There is not record of the license category: " + oldType);
+            }
 
         } catch (SQLException e) {
             e.printStackTrace();
@@ -76,12 +87,19 @@ public class TestTypeServices {
     }
 
     public void deleteTestType(String type) {
+        if (type == null) {
+            throw new NullPointerException("The test type is null");
+        }
         String sql = "DELETE FROM TestType WHERE testType = ?";
         try (Connection conn = DataBaseConnection.getConnection();
              PreparedStatement pstmt = conn.prepareStatement(sql)) {
 
             pstmt.setString(1, type);
-            pstmt.executeUpdate();
+            int affectedrows = pstmt.executeUpdate();
+            if (affectedrows == 0) {
+                throw new NoSuchElementException("There is not record of the license category: " + type);
+            }
+
 
         } catch (SQLException e) {
             e.printStackTrace();
